+++ crates/ika-core/src/dwallet_mpc/mpc_manager.rs
@@ -397,7 +397,10 @@ impl DWalletMPCManager {
     /// If no local CPUs are available, computations will execute as CPUs are freed.
     pub(crate) fn perform_cryptographic_computation(&mut self) {
         while !self.pending_for_computation_order.is_empty() {
-            if !self.cryptographic_computations_orchestrator.can_spawn_session() {
+            if !self
+                .cryptographic_computations_orchestrator
+                .can_spawn_session()
+            {
                 return;
             }
             let oldest_pending_session = self.pending_for_computation_order.pop_front().unwrap();
@@ -418,131 +421,21 @@ impl DWalletMPCManager {
                 continue;
             };
             if matches!(event_data.init_protocol_data, MPCProtocolInitData::Sign(..)) {
-                if let Err(err) = self.cryptographic_computations_orchestrator.spawn_aggregated_sign(oldest_pending_session) {
+                if let Err(err) = self
+                    .cryptographic_computations_orchestrator
+                    .spawn_aggregated_sign(oldest_pending_session)
+                {
                     error!("failed to spawn session with err: {:?}", err);
                 }
             } else {
-                if let Err(err) = self.cryptographic_computations_orchestrator.spawn_aggregated_sign(oldest_pending_session) {
+                if let Err(err) = self
+                    .cryptographic_computations_orchestrator
+                    .spawn_aggregated_sign(oldest_pending_session)
+                {
                     error!("failed to spawn session with err: {:?}", err);
                 }
             }
         }
-    }
-
-    fn spawn_session(&mut self, session: &DWalletMPCSession) -> DwalletMPCResult<()> {
-        let Some(mpc_event_data) = &session.mpc_event_data else {
-            return Err(DwalletMPCError::MissingEventDrivenData);
-        };
-        let session_id = session.session_id;
-        if self
-            .mpc_sessions
-            .get(&session_id)
-            .ok_or(DwalletMPCError::MPCSessionNotFound { session_id })?
-            .status
-            != MPCSessionStatus::Active
-        {
-            return Ok(());
-        }
-        // Hook the tokio thread pool to the rayon thread pool.
-        let handle = Handle::current();
-        let session = session.clone();
-        let finished_computation_sender = self
-            .cryptographic_computations_orchestrator
-            .computation_channel_sender
-            .clone();
-        if matches!(
-            mpc_event_data.init_protocol_data,
-            MPCProtocolInitData::Sign(..)
-        ) && session.pending_quorum_for_highest_round_number == LAST_SIGN_ROUND_INDEX
-        {
-            self.spawn_aggregated_sign(session_id, handle, session, finished_computation_sender)?;
-        } else {
-            if let Err(err) = finished_computation_sender.send(ComputationUpdate::Started) {
-                error!(
-                    "Failed to send a started computation message with error: {:?}",
-                    err
-                );
-            }
-            rayon::spawn_fifo(move || {
-                if let Err(err) = session.advance(&handle) {
-                    error!("failed to advance session with error: {:?}", err);
-                }
-                if let Err(err) = finished_computation_sender.send(ComputationUpdate::Completed) {
-                    error!(
-                        "Failed to send a finished computation message with error: {:?}",
-                        err
-                    );
-                }
-            });
-        }
-        Ok(())
-    }
-
-    fn spawn_aggregated_sign(
-        &mut self,
-        session_id: ObjectID,
-        handle: Handle,
-        ready_to_advance_session: DWalletMPCSession,
-        finished_computation_sender: UnboundedSender<ComputationUpdate>,
-    ) -> DwalletMPCResult<()> {
-        let validator_position =
-            self.get_validator_position(&ready_to_advance_session.session_id)?;
-        let epoch_store = self.epoch_store()?;
-        tokio::spawn(async move {
-            for _ in 0..validator_position {
-                let manager = epoch_store.get_dwallet_mpc_manager().await;
-                let Some(session) = manager.mpc_sessions.get(&session_id) else {
-                    error!(
-                    "failed to get session when checking if sign last round should get executed"
-                );
-                    return;
-                };
-                // If a malicious report has been received for the sign session, all the validators
-                // should execute the last step immediately.
-                if !session.session_specific_state.is_none() {
-                    break;
-                }
-                tokio::time::sleep(tokio::time::Duration::from_secs(
-                    SIGN_LAST_ROUND_COMPUTATION_CONSTANT_SECONDS as u64,
-                ))
-                .await;
-            }
-            let manager = epoch_store.get_dwallet_mpc_manager().await;
-            let Some(live_session) = manager.mpc_sessions.get(&session_id) else {
-                error!(
-                    "failed to get session when checking if sign last round should get executed"
-                );
-                return;
-            };
-            if live_session.status != MPCSessionStatus::Active
-                && !live_session.is_verifying_sign_ia_report()
-            {
-                return;
-            }
-            info!(
-                "running last sign cryptographic step for session_id: {:?}",
-                session_id
-            );
-            let session = ready_to_advance_session.clone();
-            if let Err(err) = finished_computation_sender.send(ComputationUpdate::Started) {
-                error!(
-                    "Failed to send a started computation message with error: {:?}",
-                    err
-                );
-            }
-            rayon::spawn_fifo(move || {
-                if let Err(err) = session.advance(&handle) {
-                    error!("failed to advance session with error: {:?}", err);
-                }
-                if let Err(err) = finished_computation_sender.send(ComputationUpdate::Completed) {
-                    error!(
-                        "Failed to send a finished computation message with error: {:?}",
-                        err
-                    );
-                }
-            });
-        });
-        Ok(())
     }
 
     /// Returns the epoch store.
@@ -551,24 +444,6 @@ impl DWalletMPCManager {
         self.epoch_store
             .upgrade()
             .ok_or(DwalletMPCError::EpochEnded(self.epoch_id))
-    }
-
-    /// Deterministically decides by the session ID how long this validator should wait before
-    /// running the last step of the sign protocol.
-    /// If while waiting, the validator receives a valid signature for this session,
-    /// it will not run the last step in the sign protocol, and save computation resources.
-    fn get_validator_position(&self, session_id: &ObjectID) -> DwalletMPCResult<usize> {
-        let session_id_as_32_bytes: [u8; 32] = session_id.into_bytes();
-        let positions = &self
-            .epoch_store()?
-            .committee()
-            .shuffle_by_stake_from_tx_digest(&TransactionDigest::new(session_id_as_32_bytes));
-        let authority_name = &self.epoch_store()?.name;
-        let position = positions
-            .iter()
-            .position(|&x| x == *authority_name)
-            .ok_or(DwalletMPCError::InvalidMPCPartyType)?;
-        Ok(position)
     }
 
     /// Handles a message by forwarding it to the relevant MPC session.
